+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * ajouts de types de structures liées à l'inclusion numérique par @marc-gavanier
 * nouveau champ `page_web` sur le schéma service
 * nouvelles thématiques liées au logement et à l'hébergement pour action logement
+* ajouts de nouveaux profils
 
 ### Changements